+++ packages/pwa-kit-react-sdk/src/ssr/server/express.js
@@ -118,6 +118,8 @@ export const REMOTE_REQUIRED_ENV_VARS = [
  * either as an absolute path, or relative to the build directory. If this
  * value is not supplied, requests for a favicon will return a 404 and
  * log a warning to the console.
+ * @param {Boolean} [options.loopbackAgent] - When set to true, loopback connections
+ * will be reused by means of a keep alive agent. Defaults to 'false'.
  * @param {Object} options.mobify - The 'mobify' object from the project's
  * package.json file, containing the SSR parameters.
  * @param {Number} [options.port=3443] - the localhost port on which the local
@@ -143,6 +145,9 @@ export const createApp = (options) => {
 
         // The cache time for SSR'd pages (defaults to 600 seconds)
         defaultCacheTimeSeconds: 600,
+
+        // Use a keep-alive agent for loop-back connections.
+        loopbackAgent: false,
 
         // The port that the local dev server listens on
         port: 3443,
@@ -1682,11 +1687,10 @@ const applyPatches = once((options) => {
     // Patch the http.request/get and https.request/get
     // functions to allow us to intercept them (since
     // there are multiple ways to make requests in Node).
-    const getAppHost = () => options.appHostname
-    http.request = outgoingRequestHook(http.request, getAppHost)
-    http.get = outgoingRequestHook(http.get, getAppHost)
-    https.request = outgoingRequestHook(https.request, getAppHost)
-    https.get = outgoingRequestHook(https.get, getAppHost)
+    http.request = outgoingRequestHook(http.request, options)
+    http.get = outgoingRequestHook(http.get, options)
+    https.request = outgoingRequestHook(https.request, options)
+    https.get = outgoingRequestHook(https.get, options)
 
     // Patch the ExpressJS Response class's redirect function to suppress
     // the creation of a body (DESKTOP-485). Including the body may