+++ packages/pwa-kit-react-sdk/src/utils/ssr-server.js
@@ -8,6 +8,8 @@
  * @module progressive-web-sdk/utils/ssr-server
  */
 import crypto from 'crypto'
+import http from 'http'
+import https from 'https'
 import proxy from 'http-proxy-middleware'
 import UserAgentParser from 'ua-parser-js'
 import {
@@ -167,6 +169,31 @@ export const getFullRequestURL = (url) => {
     )
 }
 
+let HTTP_AGENT, HTTPS_AGENT
+/**
+ * Returns the http and https agent singletons configured with defualt
+ * options.
+ *
+ * @private
+ * @returns {object} -
+ */
+const getAgents = () => {
+    const options = {
+        keepAlive: true,
+        keepAliveMsecs: 60 * 1000 // 1 minutes
+    }
+
+    if (!HTTP_AGENT || !HTTPS_AGENT) {
+        HTTP_AGENT = new http.Agent(options)
+        HTTPS_AGENT = new https.Agent(options)
+    }
+
+    return {
+        httpAgent: HTTP_AGENT,
+        httpsAgent: HTTPS_AGENT
+    }
+}
+
 /**
  * This function can be used to wrap http.request, http.get
  * (and the https module equivalents) in the Express app node
@@ -186,20 +213,20 @@ export const getFullRequestURL = (url) => {
  *
  * @private
  * @param wrapped {Function} the request/get function to wrap
- * @param getAppHost {Function} a function that returns the appHostname
- * for loopback requests
+ * @param options {Object} the options passed in to the servers
+ * `createApp` function.
  * @returns {Function} a function that wraps 'wrapped'
  */
 
-export const outgoingRequestHook = (wrapped, getAppHost) => {
+export const outgoingRequestHook = (wrapped, options) => {
     return function() {
         // Get the app hostname. If we can't, then just pass
         // the call through to the wrapped function. We'll also
         // do that if there's no access key.
         const accessKey = process.env.X_MOBIFY_ACCESS_KEY
-        const appHost = getAppHost && getAppHost()
+        const {appHostname, loopbackAgent} = options || {}
 
-        if (!(appHost && accessKey)) {
+        if (!(appHostname && accessKey)) {
             return wrapped.apply(this, arguments) // eslint-disable-line prefer-rest-params
         }
 
@@ -250,9 +277,9 @@ export const outgoingRequestHook = (wrapped, getAppHost) => {
             // say that 'hostname' is an alias for 'host', but that's not
             // exactly true - host can include a port but hostname doesn't
             // always. So we need to compare both.
-            workingOptions.host === appHost ||
-            workingOptions.hostname === appHost ||
-            (workingUrl && workingUrl.includes(`//${appHost}`))
+            workingOptions.host === appHostname ||
+            workingOptions.hostname === appHostname ||
+            (workingUrl && workingUrl.includes(`//${appHostname}`))
 
         if (!isLoopback) {
             return wrapped.apply(this, arguments) // eslint-disable-line prefer-rest-params
@@ -264,6 +291,22 @@ export const outgoingRequestHook = (wrapped, getAppHost) => {
 
         // Inject the access key.
         workingOptions.headers['x-mobify-access-key'] = accessKey
+
+        if (loopbackAgent) {
+            const {httpAgent, httpsAgent} = getAgents()
+
+            // Add default agent to global connection reuse.
+            workingOptions.agent =
+                workingUrl.startsWith('http:') || workingOptions?.protocol === 'http:'
+                    ? httpAgent
+                    : httpsAgent
+
+            // node-fetch and potentially other libraries add connection: close heaaders
+            // remove them to keep the connection alive. NOTE: There are variations in
+            // whether or not the connection header is upper or lower case, so handle both.
+            delete workingOptions?.headers?.connection
+            delete workingOptions?.headers?.Connection
+        }
 
         // Build the args, omitting any undefined values
         const workingArgs = [workingUrl, workingOptions, workingCallback].filter((arg) => !!arg)