+++ packages/pwa-kit-react-sdk/CHANGELOG.md
@@ -1,4 +1,7 @@
 ## v1.4.0-dev (Jan 06, 2022)
+
+-   Add `loopbackAgent` ssr-server option. [#306](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/306)
+
 ## v1.3.0 (Jan 06, 2022)
 
 -   Add `__server_only` and `__pretty_print` server rendering flags aliases. [#250](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/250)